backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-restful>=0.3.10
scikit-learn>=1.4.0
joblib>=1.3.0
nltk>=3.8.0
tldextract>=5.1.0
requests>=2.31.0
python-dotenv>=1.0.0
xgboost>=2.0.0
gunicorn>=21.2.0
